refactor: rename `db` folder to `api`
