+++ src/entities/author/api/getAuthorData.ts
