+++ src/entities/author/index.ts
@@ -1,5 +1,5 @@
-export { getAllAuthorsData } from "./db/getAllAuthorsData";
-export { getAuthorData } from "./db/getAuthorData";
+export { getAllAuthorsData } from "./api/getAllAuthorsData";
+export { getAuthorData } from "./api/getAuthorData";
 export type { AuthorData, AuthorState } from "./model/types";
 export {
   authorSlice,
