+++ src/features/profile/index.ts
@@ -1,6 +1,6 @@
 export { ProfileDetails } from "./ui/ProfileDetails/ProfileDetails";
 export { ProfileImages } from "./ui/ProfileImages/ProfileImages";
 export { ProfileUpdateForm } from "./ui/ProfileUpdateForm/ProfileUpdateForm";
-export { manageFollowStatus } from "./db/manageFollowStatus";
-export { uploadUserAvatar } from "./db/uploadUserAvatar";
+export { manageFollowStatus } from "./api/manageFollowStatus";
+export { uploadUserAvatar } from "./api/uploadUserAvatar";
 export * from "./hooks";
