+++ src/features/tweet/api/pushNewTweet.ts
@@ -1,7 +1,7 @@
 import type { TweetDBInfo, TweetType } from "@/entities/tweet";
 import { updateUserData } from "@/entities/user";
 import { updateData, uploadImage } from "@/shared/lib/firebase";
-import { pushData } from "@/shared/lib/firebase/db/pushData";
+import { pushData } from "@/shared/lib/firebase/api/pushData";
 
 export async function pushNewTweet(
   authorId: string,
