+++ src/features/tweet/index.ts
@@ -1,4 +1,4 @@
-export { pushNewTweet } from "./db/pushNewTweet";
+export { pushNewTweet } from "./api/pushNewTweet";
 export * from "./hooks";
 export { CreateTweetModalButton } from "./ui/CreateTweetModalButton/CreateTweetModalButton";
 export { TweetComposer } from "./ui/TweetComposer/TweetComposer";
