+++ src/shared/lib/firebase/api/__tests__/getChildData.test.ts
