+++ src/shared/lib/firebase/api/__tests__/getData.test.ts
