+++ src/shared/lib/firebase/api/__tests__/isDataExists.test.ts
