+++ src/shared/lib/firebase/api/__tests__/uploadImage.test.ts
