+++ src/shared/lib/firebase/api/getData.ts
