+++ src/shared/lib/firebase/api/index.ts
