+++ src/shared/lib/firebase/api/isDataExists.ts
