+++ src/shared/lib/firebase/api/uploadImage.ts
