+++ src/shared/lib/firebase/index.ts
@@ -1,2 +1,2 @@
 export * from "./app";
-export * from "./db";
+export * from "./api";
